refactor(runner): extract engine.ts and make sandbox_agent.ts a thin facade
Move shouldPark and runSandboxAgent into engine.ts. sandbox_agent.ts is now only
re-exports of the public surface, unchanged from before the split. Pure code move;
typecheck clean; 1158 runner unit tests pass.
